Fix wrong display of projects
When the user tried to create a project and click the "Create Project" button again, it didn't work because when displayProjects function was run, a new instance of "Create Project" and "Create Todo" buttons were created without the event listeners.
I fixed this by removing the code that created new instances of the buttons and instead checked if the number of child elements of ul minus 3(the two buttons and "Project" title) is greater than the current iteration, it would skip the current one and go to the next iteration without creating an already existing project.