Fix new project not showing in selection
After the user creates a project and tries to create a new to do, the user won't be able to select the new project since the populateSelectProject() function is only called when the webpage is first loaded. I fixed this by calling the function everytime a new project is created.