Fix options being repeatedly displayed
Under the select element, projects would repeatedly be displayed after a new project is created. The cause is that the populateSelectProject() function adds all the projects in the 'projects' array without checking if it is already displayed. I fixed this by removing the childs of the select element every time the populateSelectProject() function is called.