Enable todo deletion
I added a function to the 'Project' class that takes a list as an argument and then removes it from the array containing all lists. After that, I added event listeners to each delete button to pass the list that contains the delete button to the 'deleteList()' function in the Project class.